Card #226 identifier security and tests, fixes for ui, move identifier controller/service calls/tests out of user into identifier checkstyle

+++ root/api/src/main/java/org/patientview/api/service/impl/IdentifierServiceImpl.java
@@ -1,9 +1,13 @@
 package org.patientview.api.service.impl;
 
+import org.patientview.persistence.model.GroupRole;
 import org.patientview.api.service.IdentifierService;
+import org.patientview.config.exception.ResourceForbiddenException;
 import org.patientview.config.exception.ResourceNotFoundException;
 import org.patientview.persistence.model.Identifier;
+import org.patientview.persistence.model.User;
 import org.patientview.persistence.repository.IdentifierRepository;
+import org.patientview.persistence.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import javax.inject.Inject;
@@ -19,27 +23,44 @@ public class IdentifierServiceImpl extends AbstractServiceImpl<IdentifierService
     @Inject
     private IdentifierRepository identifierRepository;
 
-    public Identifier get(final Long identifierId) {
-        return identifierRepository.findOne(identifierId);
+    @Inject
+    private UserRepository userRepository;
+
+    private static final String GENERIC_GROUP_CODE = "GENERIC";
+
+    public Identifier get(final Long identifierId) throws ResourceNotFoundException, ResourceForbiddenException {
+
+        Identifier identifier = identifierRepository.findOne(identifierId);
+
+        if (identifier == null) {
+            throw new ResourceNotFoundException("Identifier does not exist");
+        }
+
+        if (!isMemberOfCurrentUsersGroups(identifier.getUser())) {
+            throw new ResourceForbiddenException("Forbidden");
+        }
+
+        return identifier;
     }
 
-    public void delete(final Long identifierId) {
+    public void delete(final Long identifierId) throws ResourceNotFoundException, ResourceForbiddenException {
+
+        Identifier identifier = identifierRepository.findOne(identifierId);
+
+        if (identifier == null) {
+            throw new ResourceNotFoundException("Identifier does not exist");
+        }
+
+        if (!isMemberOfCurrentUsersGroups(identifier.getUser())) {
+            throw new ResourceForbiddenException("Forbidden");
+        }
+
         identifierRepository.delete(identifierId);
     }
 
-    public Identifier save(final Identifier identifier) {
-        Identifier entityIdentifier = identifierRepository.findOne(identifier.getId());
-        entityIdentifier.setIdentifier(identifier.getIdentifier());
-        entityIdentifier.setIdentifierType(identifier.getIdentifierType());
-        return identifierRepository.save(entityIdentifier);
-    }
+    public void save(Identifier identifier)
+            throws ResourceNotFoundException, ResourceForbiddenException, EntityExistsException {
 
-    public Identifier add(Identifier identifier) {
-        LOG.info("Not implemented");
-        return null;
-    }
-
-    public void saveIdentifier(Identifier identifier) throws ResourceNotFoundException, EntityExistsException {
         Identifier entityIdentifier = identifierRepository.findOne(identifier.getId());
         if (entityIdentifier == null) {
             throw new ResourceNotFoundException("Identifier does not exist");
@@ -47,13 +68,65 @@ public class IdentifierServiceImpl extends AbstractServiceImpl<IdentifierService
 
         Identifier existingIdentifier = identifierRepository.findByValue(identifier.getIdentifier());
 
-        if (!existingIdentifier.equals(entityIdentifier)) {
+        if (existingIdentifier != null && !existingIdentifier.equals(entityIdentifier)) {
             throw new EntityExistsException("Cannot save Identifier, another Identifier with the same "
                     + "value already exists");
+        }
+
+        if (!isMemberOfCurrentUsersGroups(entityIdentifier.getUser())) {
+            throw new ResourceForbiddenException("Forbidden");
         }
 
         entityIdentifier.setIdentifier(identifier.getIdentifier());
         entityIdentifier.setIdentifierType(identifier.getIdentifierType());
         identifierRepository.save(entityIdentifier);
+    }
+
+    public Identifier add(Long userId, Identifier identifier)
+            throws ResourceNotFoundException, ResourceForbiddenException, EntityExistsException {
+
+        User user = userRepository.findOne(userId);
+        if (user == null) {
+            throw new ResourceNotFoundException("Could not find user");
+        }
+
+        // check Identifier doesn't already exist for another user
+        Identifier entityIdentifier = identifierRepository.findByValue(identifier.getIdentifier());
+
+        if (entityIdentifier != null) {
+            if (!(user.getId().equals(entityIdentifier.getUser().getId()))) {
+                throw new EntityExistsException("Identifier already exists for another patient");
+            }
+        }
+
+        if (!isMemberOfCurrentUsersGroups(user)) {
+            throw new ResourceForbiddenException("Forbidden");
+        }
+
+        identifier.setCreator(getCurrentUser());
+        user.getIdentifiers().add(identifier);
+        identifier.setUser(user);
+        return identifierRepository.save(identifier);
+    }
+
+    public Identifier getIdentifierByValue(String identifierValue) throws ResourceNotFoundException {
+        Identifier identifier = identifierRepository.findByValue(identifierValue);
+        if (identifier == null) {
+            throw new ResourceNotFoundException(String.format("Could not find identifier with value %s",
+                identifierValue));
+        }
+        return identifier;
+    }
+
+    private boolean isMemberOfCurrentUsersGroups(User user) {
+        for (GroupRole groupRole : user.getGroupRoles()) {
+            if (!groupRole.getGroup().getCode().toUpperCase().equals(GENERIC_GROUP_CODE)) {
+                if (isMemberOfGroup(groupRole.getGroup(), getCurrentUser())) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
     }
 }
